separate out test_get
